silence warnings introduced by add_procs merge

+++ ompi/group/group.c
@@ -557,7 +557,7 @@ int ompi_group_compare(ompi_group_t *group1,
 
 bool ompi_group_have_remote_peers (ompi_group_t *group)
 {
-    for (size_t i = 0 ; i < group->grp_proc_count ; ++i) {
+    for (int i = 0 ; i < group->grp_proc_count ; ++i) {
         ompi_proc_t *proc = NULL;
 #if OMPI_GROUP_SPARSE
         proc = ompi_group_peer_lookup (group, i);
